Note local tag xf86-input-fpit-1.1.0@mandriva suggested on upstream  Tag at git checkout 47ecabff271fc1b8dfcc40656934fb70264b7a0e
  Notes about this package:

  Currently cooker has an x11-driver-input-fpit at version 1.1.0
  There isn't a 1.1.1 tag, but probably this package should have it's version
updated.
  There is a fpit-1_1_0 tag upstream, that matches 2008 and current cooker
version, i.e. the only "broken" symbol is xf86IsCorePointer, but since based
on the logs the adition of RandR support seens logical and can be disabled
with a proper option, it is currently also broken due to using older xrandr
version, and even in the comments it says that probably it would be better
to use another function/interface.


SILENT: svn-revision: 132431